pyproject.toml
[project]
name = "nonebot-plugin-flo-luck"
version = "0.2.0"
description = "nonebot plugin for today's luck"
readme = "README.md"
requires-python = ">=3.9, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_alconna", "nonebot_plugin_uninfo", "nonebot_plugin_localstore", "nonebot_plugin_flo_luck"]
plugin_dirs = []
builtin_plugins = []

[tool.poetry]
name = "nonebot-plugin-flo-luck"
version = "0.2.0"
description = "nonebot plugin for today's luck."
authors = ["florenz <florenz@tju.edu.cn>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = ">=2.4.0"
nonebot-plugin-alconna = ">=0.54.1"
nonebot-plugin-uninfo = ">=0.6.2"
nonebot-plugin-localstore = ">=0.7.2"
